Reorg rank_eval
