+++ specification/_global/rank_eval/types.ts
@@ -17,7 +17,11 @@
  * under the License.
  */
 
-import { integer } from '@_types/Numeric'
+import { Dictionary } from '@spec_utils/Dictionary'
+import { UserDefinedValue } from '@spec_utils/UserDefinedValue'
+import { Id, IndexName, Type } from '@_types/common'
+import { double, integer } from '@_types/Numeric'
+import { QueryContainer } from '@_types/query_dsl/abstractions'
 
 export class RankEvalMetricBase {
   /**
@@ -89,4 +93,59 @@ export class RankEvalMetric {
   mean_reciprocal_rank?: RankEvalMetricMeanReciprocalRank
   dcg?: RankEvalMetricDiscountedCumulativeGain
   expected_reciprocal_rank?: RankEvalMetricExpectedReciprocalRank
+}
+
+export class RankEvalRequestItem {
+  /** The search request’s ID, used to group result details later. */
+  id: Id
+  /** The query being evaluated. */
+  request?: RankEvalQuery
+  /** List of document ratings */
+  ratings: DocumentRating[]
+  /** The search template Id */
+  template_id?: Id
+  /** The search template parameters. */
+  params?: Dictionary<string, UserDefinedValue>
+}
+
+export class RankEvalQuery {
+  query: QueryContainer
+  size?: integer
+}
+
+export class DocumentRating {
+  /** The document ID. */
+  _id: Id
+  /** The document’s index. For data streams, this should be the document’s backing index. */
+  _index: IndexName
+  /** The document’s relevance with regard to this search request. */
+  rating: integer
+}
+
+export class RankEvalMetricDetail {
+  /** The metric_score in the details section shows the contribution of this query to the global quality metric score */
+  metric_score: double
+  /** The unrated_docs section contains an _index and _id entry for each document in the search result for this query that didn’t have a ratings value. This can be used to ask the user to supply ratings for these documents */
+  unrated_docs: UnratedDocument[]
+  /** The hits section shows a grouping of the search results with their supplied ratings */
+  hits: RankEvalHitItem[]
+  /** The metric_details give additional information about the calculated quality metric (e.g. how many of the retrieved documents were relevant). The content varies for each metric but allows for better interpretation of the results */
+  metric_details: Dictionary<string, Dictionary<string, UserDefinedValue>>
+}
+
+export class RankEvalHitItem {
+  hit: RankEvalHit
+  rating?: double
+}
+
+export class RankEvalHit {
+  _id: Id
+  _index: IndexName
+  _type?: Type
+  _score: double
+}
+
+export class UnratedDocument {
+  _id: Id
+  _index: IndexName
 }
